pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "livekit-agents>=1.3.11",
    "livekit-plugins-cartesia>=1.3.11",
    "livekit-plugins-deepgram>=1.3.11",
    "livekit-plugins-google>=1.3.11",
    "livekit-plugins-openai>=1.3.11",
    "livekit-plugins-silero>=1.3.11",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
]
